Fixed Multiplayer MouseSynicng

+++ Content/CursedTechniques/BloodManipulation/BloodDaggerStormProjectile.cs
@@ -51,30 +51,31 @@ namespace sorceryFight.Content.CursedTechniques.BloodManipulation
             }
 
             //Main.NewText("Target State" + Projectile.ai[1]);
-
-            // Validate target first
-            if (Projectile.ai[1] < 0 || !Main.npc[(int)Projectile.ai[1]].active)
-                Projectile.ai[1] = FindTarget();
-
-            if (Projectile.ai[1] >= 0 && Main.npc[(int)Projectile.ai[1]].Distance(Projectile.Center) < trackingRadius)
+            if (Main.myPlayer == Projectile.owner)
             {
-                Vector2 targetVelocity = (Main.npc[(int)Projectile.ai[1]].Center - Projectile.Center).SafeNormalize(Vector2.Zero) * 20f;
-                Projectile.velocity = Vector2.Lerp(Projectile.velocity, targetVelocity, 0.25f);
-            }
-            else
-            {
-                Vector2 mouseVelocity = (Main.MouseWorld - Projectile.Center).SafeNormalize(Vector2.Zero) * 20f;
-                Projectile.velocity = Vector2.Lerp(Projectile.velocity, mouseVelocity, 0.25f);
-                noTargetCounter++;
-            }
+                // Validate target first
+                if (Projectile.ai[1] < 0 || !Main.npc[(int)Projectile.ai[1]].active)
+                    Projectile.ai[1] = FindTarget();
 
-            if (160 < noTargetCounter)
-            {
-                Projectile.Kill();
+                if (Projectile.ai[1] >= 0 && Main.npc[(int)Projectile.ai[1]].Distance(Projectile.Center) < trackingRadius)
+                {
+                    Vector2 targetVelocity = (Main.npc[(int)Projectile.ai[1]].Center - Projectile.Center).SafeNormalize(Vector2.Zero) * 20f;
+                    Projectile.velocity = Vector2.Lerp(Projectile.velocity, targetVelocity, 0.25f);
+                }
+                else
+                {
+                    Vector2 mouseVelocity = (Main.MouseWorld - Projectile.Center).SafeNormalize(Vector2.Zero) * 20f;
+                    Projectile.velocity = Vector2.Lerp(Projectile.velocity, mouseVelocity, 0.25f);
+                    noTargetCounter++;
+                }
+
+                if (160 < noTargetCounter)
+                {
+                    Projectile.Kill();
+                }
+
+                Projectile.rotation = Projectile.velocity.ToRotation();
             }
-
-            Projectile.rotation = Projectile.velocity.ToRotation();
-
             Vector2 behindOffset = -Projectile.velocity.SafeNormalize(Vector2.Zero) * Main.rand.NextFloat(10f, 40f);
             Vector2 particleOffset = Projectile.Center + behindOffset;
             Vector2 particleVelocity = particleOffset.DirectionTo(Projectile.Center);
